+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -126,7 +126,6 @@ export enum TelemetryEvent {
     ErrorUIPaneLoaded = "ErrorUIPaneLoaded",
     DownloadTranscriptFailed = "DownloadTranscriptFailed",
     StartChatFailed = "StartChatFailed",
-    IC3ThreadUpdateEventReceived = "IC3ThreadUpdateEventReceived",
     ConfirmationCancelButtonClicked = "ConfirmationCancelButtonClicked",
     ConfirmationConfirmButtonClicked = "ConfirmationConfirmButtonClicked",
     LoadingPaneLoaded = "LoadingPaneLoaded",
@@ -264,7 +263,6 @@ export class TelemetryConstants {
             case TelemetryEvent.EmailTranscriptSent:
             case TelemetryEvent.EmailTranscriptFailed:
             case TelemetryEvent.DownloadTranscriptFailed:
-            case TelemetryEvent.IC3ThreadUpdateEventReceived:
             case TelemetryEvent.ConfirmationCancelButtonClicked:
             case TelemetryEvent.ConfirmationConfirmButtonClicked:
             case TelemetryEvent.PreChatSurveyStartChatMethodFailed: